requirements.txt
groq>=0.9.0
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.20.0
matplotlib>=3.8.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
scipy>=1.12.0
